IMMORTAL: Util -> Utilities, and addition of addSprite to utilities
